# Align allocations