For AMD64 aligned inputs can make a big speed difference.

This is an example of the speed difference when inputs are unaligned/aligned:

```
BenchmarkEncode100x20x10000-32    	    7058	    172648 ns/op	6950.57 MB/s
BenchmarkEncode100x20x10000-32    	    8406	    137911 ns/op	8701.24 MB/s
```

This is mostly the case when dealing with odd-sized shards. 

To facilitate this the package provides an `AllocAligned(shards, each int) [][]byte`. 
This will allocate a number of shards, each with the size `each`.
Each shard will then be aligned to a 64 byte boundary.

Each encoder also has a `AllocAligned(each int) [][]byte` as an extended interface which will return the same, 
but with the shard count configured in the encoder.   

It is not possible to re-aligned already allocated slices, for example when using `Split`.
When it is not possible to write to aligned shards, you should not copy to them.

Full (but rather noisy) benchmark:

```
benchmark                                                old ns/op      new ns/op      delta
BenchmarkGalois128K-32                                   2284           2254           -1.31%
BenchmarkGalois1M-32                                     21925          19042          -13.15%
BenchmarkGaloisXor128K-32                                2810           2782           -1.00%
BenchmarkGaloisXor1M-32                                  24223          22716          -6.22%
BenchmarkEncode2x1x1M-32                                 38969          33115          -15.02%
BenchmarkEncode800x200/64-32                             29007          28090          -3.16%
BenchmarkEncode800x200/256-32                            65858          64747          -1.69%
BenchmarkEncode800x200/1024-32                           207661         203905         -1.81%
BenchmarkEncode800x200/4096-32                           806579         789913         -2.07%
BenchmarkEncode800x200/16384-32                          4088967        3688426        -9.80%
BenchmarkEncode800x200/65536-32                          27241951       24104804       -11.52%
BenchmarkEncode800x200/262144-32                         120608789      113648633      -5.77%
BenchmarkEncode800x200/1048576-32                        451364367      420720500      -6.79%
BenchmarkEncode1K/4+4/cauchy-32                          335            345            +2.96%
BenchmarkEncode1K/4+4/leopard-gf8-32                     640            632            -1.33%
BenchmarkEncode1K/4+4/leopard-gf16-32                    455            436            -4.16%
BenchmarkEncode1K/8+8/cauchy-32                          1099           1081           -1.64%
BenchmarkEncode1K/8+8/leopard-gf8-32                     1831           1792           -2.13%
BenchmarkEncode1K/8+8/leopard-gf16-32                    1608           1586           -1.37%
BenchmarkEncode1K/16+16/cauchy-32                        4340           4372           +0.74%
BenchmarkEncode1K/16+16/leopard-gf8-32                   3330           3280           -1.50%
BenchmarkEncode1K/16+16/leopard-gf16-32                  2637           2614           -0.87%
BenchmarkEncode1K/32+32/cauchy-32                        17257          17397          +0.81%
BenchmarkEncode1K/32+32/leopard-gf8-32                   9849           9623           -2.29%
BenchmarkEncode1K/32+32/leopard-gf16-32                  8903           8806           -1.09%
BenchmarkEncode1K/64+64/cauchy-32                        68672          68374          -0.43%
BenchmarkEncode1K/64+64/leopard-gf8-32                   18283          17992          -1.59%
BenchmarkEncode1K/64+64/leopard-gf16-32                  15558          15541          -0.11%
BenchmarkEncode1K/128+128/cauchy-32                      270881         270547         -0.12%
BenchmarkEncode1K/128+128/leopard-gf8-32                 49601          48871          -1.47%
BenchmarkEncode1K/128+128/leopard-gf16-32                46158          45735          -0.92%
BenchmarkEncode1K/256+256/leopard-gf16-32                84268          83318          -1.13%
BenchmarkEncode1K/512+512/leopard-gf16-32                235278         231775         -1.49%
BenchmarkEncode1K/1024+1024/leopard-gf16-32              436245         430979         -1.21%
BenchmarkEncode1K/2048+2048/leopard-gf16-32              1227665        1108337        -9.72%
BenchmarkEncode1K/4096+4096/leopard-gf16-32              2573166        2273580        -11.64%
BenchmarkEncode1K/8192+8192/leopard-gf16-32              7377235        6443287        -12.66%
BenchmarkEncode1K/16384+16384/leopard-gf16-32            20045895       17426286       -13.07%
BenchmarkEncode1K/32768+32768/leopard-gf16-32            53570005       50222577       -6.25%
BenchmarkDecode1K/4+4/cauchy-32                          2160           2262           +4.72%
BenchmarkDecode1K/4+4/cauchy-inv-32                      1369           1461           +6.72%
BenchmarkDecode1K/4+4/cauchy-single-32                   1300           1304           +0.31%
BenchmarkDecode1K/4+4/cauchy-single-inv-32               604            633            +4.71%
BenchmarkDecode1K/4+4/leopard-gf8-32                     4277           4176           -2.36%
BenchmarkDecode1K/4+4/leopard-gf8-inv-32                 2455           2330           -5.09%
BenchmarkDecode1K/4+4/leopard-gf8-single-32              3891           3750           -3.62%
BenchmarkDecode1K/4+4/leopard-gf8-single-inv-32          1974           1914           -3.04%
BenchmarkDecode1K/4+4/leopard-gf16-32                    794838         792366         -0.31%
BenchmarkDecode1K/4+4/leopard-gf16-single-32             791991         793335         +0.17%
BenchmarkDecode1K/8+8/cauchy-32                          5445           5651           +3.78%
BenchmarkDecode1K/8+8/cauchy-inv-32                      2920           3039           +4.08%
BenchmarkDecode1K/8+8/cauchy-single-32                   2290           2285           -0.22%
BenchmarkDecode1K/8+8/cauchy-single-inv-32               788            814            +3.30%
BenchmarkDecode1K/8+8/leopard-gf8-32                     7270           7134           -1.87%
BenchmarkDecode1K/8+8/leopard-gf8-inv-32                 5433           5156           -5.10%
BenchmarkDecode1K/8+8/leopard-gf8-single-32              6168           6123           -0.73%
BenchmarkDecode1K/8+8/leopard-gf8-single-inv-32          4302           4365           +1.46%
BenchmarkDecode1K/8+8/leopard-gf16-32                    792261         787290         -0.63%
BenchmarkDecode1K/8+8/leopard-gf16-single-32             800258         793031         -0.90%
BenchmarkDecode1K/16+16/cauchy-32                        21492          21664          +0.80%
BenchmarkDecode1K/16+16/cauchy-inv-32                    7805           8140           +4.29%
BenchmarkDecode1K/16+16/cauchy-single-32                 5176           5105           -1.37%
BenchmarkDecode1K/16+16/cauchy-single-inv-32             1211           1194           -1.40%
BenchmarkDecode1K/16+16/leopard-gf8-32                   16062          15813          -1.55%
BenchmarkDecode1K/16+16/leopard-gf8-inv-32               14737          13697          -7.06%
BenchmarkDecode1K/16+16/leopard-gf8-single-32            13899          13393          -3.64%
BenchmarkDecode1K/16+16/leopard-gf8-single-inv-32        11865          11487          -3.19%
BenchmarkDecode1K/16+16/leopard-gf16-32                  805833         794984         -1.35%
BenchmarkDecode1K/16+16/leopard-gf16-single-32           808125         799816         -1.03%
BenchmarkDecode1K/32+32/cauchy-32                        117646         117088         -0.47%
BenchmarkDecode1K/32+32/cauchy-inv-32                    23975          24454          +2.00%
BenchmarkDecode1K/32+32/cauchy-single-32                 14250          13861          -2.73%
BenchmarkDecode1K/32+32/cauchy-single-inv-32             2018           1994           -1.19%
BenchmarkDecode1K/32+32/leopard-gf8-32                   31886          32203          +0.99%
BenchmarkDecode1K/32+32/leopard-gf8-inv-32               30959          30270          -2.23%
BenchmarkDecode1K/32+32/leopard-gf8-single-32            20726          22461          +8.37%
BenchmarkDecode1K/32+32/leopard-gf8-single-inv-32        20717          22592          +9.05%
BenchmarkDecode1K/32+32/leopard-gf16-32                  823015         808697         -1.74%
BenchmarkDecode1K/32+32/leopard-gf16-single-32           816765         809475         -0.89%
BenchmarkDecode1K/64+64/cauchy-32                        813174         804646         -1.05%
BenchmarkDecode1K/64+64/cauchy-inv-32                    80979          82487          +1.86%
BenchmarkDecode1K/64+64/cauchy-single-32                 45979          44580          -3.04%
BenchmarkDecode1K/64+64/cauchy-single-inv-32             3432           3286           -4.25%
BenchmarkDecode1K/64+64/leopard-gf8-32                   78343          74364          -5.08%
BenchmarkDecode1K/64+64/leopard-gf8-inv-32               73370          63120          -13.97%
BenchmarkDecode1K/64+64/leopard-gf8-single-32            50130          43959          -12.31%
BenchmarkDecode1K/64+64/leopard-gf8-single-inv-32        51308          41208          -19.69%
BenchmarkDecode1K/64+64/leopard-gf16-32                  864012         846280         -2.05%
BenchmarkDecode1K/64+64/leopard-gf16-single-32           850149         830551         -2.31%
BenchmarkDecode1K/128+128/cauchy-32                      5929095        5900026        -0.49%
BenchmarkDecode1K/128+128/cauchy-inv-32                  304087         306298         +0.73%
BenchmarkDecode1K/128+128/cauchy-single-32               164090         160974         -1.90%
BenchmarkDecode1K/128+128/cauchy-single-inv-32           5850           5625           -3.85%
BenchmarkDecode1K/128+128/leopard-gf8-32                 158429         145910         -7.90%
BenchmarkDecode1K/128+128/leopard-gf8-inv-32             152309         141621         -7.02%
BenchmarkDecode1K/128+128/leopard-gf8-single-32          112267         94170          -16.12%
BenchmarkDecode1K/128+128/leopard-gf8-single-inv-32      104646         96275          -8.00%
BenchmarkDecode1K/128+128/leopard-gf16-32                927823         920083         -0.83%
BenchmarkDecode1K/128+128/leopard-gf16-single-32         893019         885971         -0.79%
BenchmarkDecode1K/256+256/leopard-gf16-32                1132479        1105774        -2.36%
BenchmarkDecode1K/256+256/leopard-gf16-single-32         1017945        1003342        -1.43%
BenchmarkDecode1K/512+512/leopard-gf16-32                1495247        1457558        -2.52%
BenchmarkDecode1K/512+512/leopard-gf16-single-32         1276089        1239965        -2.83%
BenchmarkDecode1K/1024+1024/leopard-gf16-32              2511310        2355124        -6.22%
BenchmarkDecode1K/1024+1024/leopard-gf16-single-32       1926875        1786114        -7.31%
BenchmarkDecode1K/2048+2048/leopard-gf16-32              4574758        4051357        -11.44%
BenchmarkDecode1K/2048+2048/leopard-gf16-single-32       3404487        2936912        -13.73%
BenchmarkDecode1K/4096+4096/leopard-gf16-32              9917650        9381317        -5.41%
BenchmarkDecode1K/4096+4096/leopard-gf16-single-32       7439868        6237255        -16.16%
BenchmarkDecode1K/8192+8192/leopard-gf16-32              27173871       22125130       -18.58%
BenchmarkDecode1K/8192+8192/leopard-gf16-single-32       19590423       15888578       -18.90%
BenchmarkDecode1K/16384+16384/leopard-gf16-32            65490106       60630937       -7.42%
BenchmarkDecode1K/16384+16384/leopard-gf16-single-32     43015162       40455732       -5.95%
BenchmarkDecode1K/32768+32768/leopard-gf16-32            137665712      121400489      -11.82%
BenchmarkDecode1K/32768+32768/leopard-gf16-single-32     89620746       84439785       -5.78%
BenchmarkEncodeLeopard/83840-32                          38754790       34248175       -11.63%
BenchmarkEncode10x2x10000-32                             3146           3104           -1.34%
BenchmarkEncode100x20x10000-32                           156083         132442         -15.15%
BenchmarkEncode17x3x1M-32                                266619         549537         +106.11%
BenchmarkEncode10x4x16M-32                               8589080        10118010       +17.80%
BenchmarkEncode5x2x1M-32                                 69710          67893          -2.61%
BenchmarkEncode10x2x1M-32                                105670         107245         +1.49%
BenchmarkEncode10x4x1M-32                                163118         176928         +8.47%
BenchmarkEncode50x20x1M-32                               3118381        14567976       +367.16%
BenchmarkEncodeLeopard50x20x1M-32                        10595922       11908254       +12.39%
BenchmarkEncode17x3x16M-32                               10172963       11955679       +17.52%
BenchmarkEncode_8x4x8M-32                                3728191        3439514        -7.74%
BenchmarkEncode_12x4x12M-32                              6803056        6497932        -4.49%
BenchmarkEncode_16x4x16M-32                              10882620       10934306       +0.47%
BenchmarkEncode_16x4x32M-32                              21758418       21431559       -1.50%
BenchmarkEncode_16x4x64M-32                              45258777       43288619       -4.35%
BenchmarkEncode_8x5x8M-32                                4385065        4089020        -6.75%
BenchmarkEncode_8x6x8M-32                                4826290        4502941        -6.70%
BenchmarkEncode_8x7x8M-32                                5403098        4968727        -8.04%
BenchmarkEncode_8x9x8M-32                                6302487        6002242        -4.76%
BenchmarkEncode_8x10x8M-32                               6913882        6637816        -3.99%
BenchmarkEncode_8x11x8M-32                               7326503        7059232        -3.65%
BenchmarkEncode_8x8x05M-32                               144080         311435         +116.15%
BenchmarkEncode_8x8x1M-32                                291361         274498         -5.79%
BenchmarkEncode_8x8x8M-32                                5802282        5384482        -7.20%
BenchmarkEncode_8x8x32M-32                               24160839       23571850       -2.44%
BenchmarkEncode_24x8x24M-32                              36720647       30780145       -16.18%
BenchmarkEncode_24x8x48M-32                              63394650       63509811       +0.18%
BenchmarkVerify800x200/64-32                             40375          38159          -5.49%
BenchmarkVerify800x200/256-32                            83771          78225          -6.62%
BenchmarkVerify800x200/1024-32                           271752         240426         -11.53%
BenchmarkVerify800x200/4096-32                           1082853        929348         -14.18%
BenchmarkVerify800x200/16384-32                          5725732        4615986        -19.38%
BenchmarkVerify800x200/65536-32                          32934571       26172560       -20.53%
BenchmarkVerify800x200/262144-32                         155253386      126563067      -18.48%
BenchmarkVerify800x200/1048576-32                        561659750      490361633      -12.69%
BenchmarkVerify10x2x10000-32                             5904           5500           -6.84%
BenchmarkVerify50x5x100000-32                            211479         171168         -19.06%
BenchmarkVerify10x2x1M-32                                504104         425526         -15.59%
BenchmarkVerify5x2x1M-32                                 395883         335088         -15.36%
BenchmarkVerify10x4x1M-32                                1089718        938154         -13.91%
BenchmarkVerify50x20x1M-32                               7889230        8431484        +6.87%
BenchmarkVerify10x4x16M-32                               23219342       17756713       -23.53%
BenchmarkReconstruct10x2x10000-32                        3222           3117           -3.26%
BenchmarkReconstruct800x200/64-32                        1179363        1100846        -6.66%
BenchmarkReconstruct800x200/256-32                       1294499        1214204        -6.20%
BenchmarkReconstruct800x200/1024-32                      1881303        1737606        -7.64%
BenchmarkReconstruct800x200/4096-32                      4420506        3860332        -12.67%
BenchmarkReconstruct800x200/16384-32                     27171535       22810506       -16.05%
BenchmarkReconstruct800x200/65536-32                     119789411      108652470      -9.30%
BenchmarkReconstruct800x200/262144-32                    531271300      505694550      -4.81%
BenchmarkReconstruct800x200/1048576-32                   2918642700     2408563700     -17.48%
BenchmarkReconstruct50x5x50000-32                        148109         134759         -9.01%
BenchmarkReconstruct10x2x1M-32                           185262         186595         +0.72%
BenchmarkReconstruct5x2x1M-32                            129164         126902         -1.75%
BenchmarkReconstruct10x4x1M-32                           276915         268351         -3.09%
BenchmarkReconstruct50x20x1M-32                          3541594        6128024        +73.03%
BenchmarkReconstructLeopard50x20x1M-32                   30241625       29009541       -4.07%
BenchmarkReconstruct10x4x16M-32                          9437829        8883464        -5.87%
BenchmarkReconstructData10x2x10000-32                    3060           2988           -2.35%
BenchmarkReconstructData800x200/64-32                    1147537        1091256        -4.90%
BenchmarkReconstructData800x200/256-32                   1263484        1212899        -4.00%
BenchmarkReconstructData800x200/1024-32                  1839242        1725180        -6.20%
BenchmarkReconstructData800x200/4096-32                  4362936        3849980        -11.76%
BenchmarkReconstructData800x200/16384-32                 26168836       22671938       -13.36%
BenchmarkReconstructData800x200/65536-32                 120528889      107130140      -11.12%
BenchmarkReconstructData800x200/262144-32                563569000      475725700      -15.59%
BenchmarkReconstructData800x200/1048576-32               2615405500     2429356800     -7.11%
BenchmarkReconstructData50x5x50000-32                    143226         129456         -9.61%
BenchmarkReconstructData10x2x1M-32                       172211         237621         +37.98%
BenchmarkReconstructData5x2x1M-32                        112454         100520         -10.61%
BenchmarkReconstructData10x4x1M-32                       228400         304907         +33.50%
BenchmarkReconstructData50x20x1M-32                      2314132        4094795        +76.95%
BenchmarkReconstructData10x4x16M-32                      7298997        6668680        -8.64%
BenchmarkReconstructP10x2x10000-32                       826            829            +0.35%
BenchmarkReconstructP10x5x20000-32                       1474           1574           +6.78%
BenchmarkSplit10x4x160M-32                               5724818        5133214        -10.33%
BenchmarkSplit5x2x5M-32                                  185120         158360         -14.46%
BenchmarkSplit10x2x1M-32                                 30804          26726          -13.24%
BenchmarkSplit10x4x10M-32                                362056         329514         -8.99%
BenchmarkSplit50x20x50M-32                               1822737        1658791        -8.99%
BenchmarkSplit17x3x272M-32                               4211272        3692685        -12.31%
BenchmarkParallel_8x8x64K-32                             6069           9427           +55.33%
…